Interrogate controller in Game scene for input
Previously we responded to events from the event pump, but events are
only generated when an input changes. As a result, we did not always
receive events for both axis of a thumbstick, and had to keep state in
order to figure out the complete, current orientation of a thumbstick.
But even still, we could miss some events, especially when the
thumbstick was zeroed after a quick flick, which would cause the player
to accelerate non-stop even though they gave no input.

Now we interrogate the state of the controller each frame, which gives
us the current axis readings and held buttons. This eliminates the
stateful logic, lets the player hold a button to fire, and appears to be
completley accurate.